+++ packages/anoncreds-rs/src/services/__tests__/AnonCredsRsServices.test.ts
@@ -25,7 +25,6 @@ import { anoncreds } from '@hyperledger/anoncreds-nodejs'
 import { Subject } from 'rxjs'
 
 import { InMemoryStorageService } from '../../../../../tests/InMemoryStorageService'
-import { describeRunInNodeVersion } from '../../../../../tests/runInVersion'
 import { encodeCredentialValue } from '../../../../anoncreds/src/utils/credential'
 import { InMemoryAnonCredsRegistry } from '../../../../anoncreds/tests/InMemoryAnonCredsRegistry'
 import { agentDependencies, getAgentConfig, getAgentContext } from '../../../../core/tests/helpers'
@@ -58,8 +57,7 @@ const agentContext = getAgentContext({
   agentConfig,
 })
 
-// FIXME: Re-include in tests when NodeJS wrapper performance is improved
-describeRunInNodeVersion([18], 'AnonCredsRsServices', () => {
+describe('AnonCredsRsServices', () => {
   test('issuance flow without revocation', async () => {
     const issuerId = 'did:indy:pool:localtest:TL1EaPFCZ8Si5aUrqScBDt'