+++ packages/anoncreds/src/utils/getRevocationRegistries.ts
@@ -45,6 +45,7 @@ export async function getRevocationRegistriesForRequest(
       })
     }
 
+    const revocationRegistryPromises = []
     for (const { referent, selectedCredential, nonRevoked, type } of referentCredentials) {
       if (!selectedCredential.credentialInfo) {
         throw new AriesFrameworkError(
@@ -76,30 +77,34 @@ export async function getRevocationRegistriesForRequest(
           .resolve(AnonCredsRegistryService)
           .getRegistryForIdentifier(agentContext, revocationRegistryId)
 
-        // Fetch revocation registry definition if not in revocation registries list yet
-        if (!revocationRegistries[revocationRegistryId]) {
-          const { revocationRegistryDefinition, resolutionMetadata } = await registry.getRevocationRegistryDefinition(
-            agentContext,
-            revocationRegistryId
-          )
-          if (!revocationRegistryDefinition) {
-            throw new AriesFrameworkError(
-              `Could not retrieve revocation registry definition for revocation registry ${revocationRegistryId}: ${resolutionMetadata.message}`
+        const getRevocationRegistry = async () => {
+          // Fetch revocation registry definition if not in revocation registries list yet
+          if (!revocationRegistries[revocationRegistryId]) {
+            const { revocationRegistryDefinition, resolutionMetadata } = await registry.getRevocationRegistryDefinition(
+              agentContext,
+              revocationRegistryId
             )
-          }
+            if (!revocationRegistryDefinition) {
+              throw new AriesFrameworkError(
+                `Could not retrieve revocation registry definition for revocation registry ${revocationRegistryId}: ${resolutionMetadata.message}`
+              )
+            }
 
-          const tailsFileService = agentContext.dependencyManager.resolve(AnonCredsModuleConfig).tailsFileService
-          const tailsFilePath = await tailsFileService.getTailsFile(agentContext, {
-            revocationRegistryDefinition,
-          })
+            const tailsFileService = agentContext.dependencyManager.resolve(AnonCredsModuleConfig).tailsFileService
+            const tailsFilePath = await tailsFileService.getTailsFile(agentContext, {
+              revocationRegistryDefinition,
+            })
 
-          // const tails = await this.indyUtilitiesService.downloadTails(tailsHash, tailsLocation)
-          revocationRegistries[revocationRegistryId] = {
-            definition: revocationRegistryDefinition,
-            tailsFilePath,
-            revocationStatusLists: {},
+            // const tails = await this.indyUtilitiesService.downloadTails(tailsHash, tailsLocation)
+            revocationRegistries[revocationRegistryId] = {
+              definition: revocationRegistryDefinition,
+              tailsFilePath,
+              revocationStatusLists: {},
+            }
           }
         }
+
+        revocationRegistryPromises.push(getRevocationRegistry())
 
         // In most cases we will have a timestamp, but if it's not defined, we use the nonRevoked.to value
         const timestampToFetch = timestamp ?? nonRevoked.to
@@ -135,7 +140,8 @@ export async function getRevocationRegistriesForRequest(
         }
       }
     }
-
+    // await all revocation registry statuses asynchronously
+    await Promise.all(revocationRegistryPromises)
     agentContext.config.logger.debug(`Retrieved revocation registries for proof request`, {
       revocationRegistries,
     })
@@ -155,6 +161,7 @@ export async function getRevocationRegistriesForRequest(
 export async function getRevocationRegistriesForProof(agentContext: AgentContext, proof: AnonCredsProof) {
   const revocationRegistries: VerifyProofOptions['revocationRegistries'] = {}
 
+  const revocationRegistryPromises = []
   for (const identifier of proof.identifiers) {
     const revocationRegistryId = identifier.rev_reg_id
     const timestamp = identifier.timestamp
@@ -166,38 +173,41 @@ export async function getRevocationRegistriesForProof(agentContext: AgentContext
       .resolve(AnonCredsRegistryService)
       .getRegistryForIdentifier(agentContext, revocationRegistryId)
 
-    // Fetch revocation registry definition if not already fetched
-    if (!revocationRegistries[revocationRegistryId]) {
-      const { revocationRegistryDefinition, resolutionMetadata } = await registry.getRevocationRegistryDefinition(
-        agentContext,
-        revocationRegistryId
-      )
-      if (!revocationRegistryDefinition) {
-        throw new AriesFrameworkError(
-          `Could not retrieve revocation registry definition for revocation registry ${revocationRegistryId}: ${resolutionMetadata.message}`
+    const getRevocationRegistry = async () => {
+      // Fetch revocation registry definition if not already fetched
+      if (!revocationRegistries[revocationRegistryId]) {
+        const { revocationRegistryDefinition, resolutionMetadata } = await registry.getRevocationRegistryDefinition(
+          agentContext,
+          revocationRegistryId
         )
+        if (!revocationRegistryDefinition) {
+          throw new AriesFrameworkError(
+            `Could not retrieve revocation registry definition for revocation registry ${revocationRegistryId}: ${resolutionMetadata.message}`
+          )
+        }
+
+        revocationRegistries[revocationRegistryId] = {
+          definition: revocationRegistryDefinition,
+          revocationStatusLists: {},
+        }
       }
 
-      revocationRegistries[revocationRegistryId] = {
-        definition: revocationRegistryDefinition,
-        revocationStatusLists: {},
+      // Fetch revocation status list by timestamp if not already fetched
+      if (!revocationRegistries[revocationRegistryId].revocationStatusLists[timestamp]) {
+        const { revocationStatusList, resolutionMetadata: statusListResolutionMetadata } =
+          await registry.getRevocationStatusList(agentContext, revocationRegistryId, timestamp)
+
+        if (!revocationStatusList) {
+          throw new AriesFrameworkError(
+            `Could not retrieve revocation status list for revocation registry ${revocationRegistryId}: ${statusListResolutionMetadata.message}`
+          )
+        }
+
+        revocationRegistries[revocationRegistryId].revocationStatusLists[timestamp] = revocationStatusList
       }
     }
-
-    // Fetch revocation status list by timestamp if not already fetched
-    if (!revocationRegistries[revocationRegistryId].revocationStatusLists[timestamp]) {
-      const { revocationStatusList, resolutionMetadata: statusListResolutionMetadata } =
-        await registry.getRevocationStatusList(agentContext, revocationRegistryId, timestamp)
-
-      if (!revocationStatusList) {
-        throw new AriesFrameworkError(
-          `Could not retrieve revocation status list for revocation registry ${revocationRegistryId} at ${timestamp}: ${statusListResolutionMetadata.message}`
-        )
-      }
-
-      revocationRegistries[revocationRegistryId].revocationStatusLists[timestamp] = revocationStatusList
-    }
+    revocationRegistryPromises.push(getRevocationRegistry())
   }
-
+  await Promise.all(revocationRegistryPromises)
   return revocationRegistries
 }