+++ packages/core/src/modules/oob/helpers.ts
@@ -4,7 +4,7 @@ import { ConnectionInvitationMessage, HandshakeProtocol } from '../connections'
 import { didKeyToVerkey, verkeyToDidKey } from '../dids/helpers'
 
 import { OutOfBandDidCommService } from './domain/OutOfBandDidCommService'
-import { OutOfBandInvitation } from './messages'
+import { InvitationType, OutOfBandInvitation } from './messages'
 
 export function convertToNewInvitation(oldInvitation: ConnectionInvitationMessage) {
   let service
@@ -32,7 +32,9 @@ export function convertToNewInvitation(oldInvitation: ConnectionInvitationMessag
     handshakeProtocols: [HandshakeProtocol.Connections],
   }
 
-  return new OutOfBandInvitation(options)
+  const outOfBandInvitation = new OutOfBandInvitation(options)
+  outOfBandInvitation.invitationType = InvitationType.Connection
+  return outOfBandInvitation
 }
 
 export function convertToOldInvitation(newInvitation: OutOfBandInvitation) {