+++ packages/core/src/wallet/error/WalletImportPathExistsError.ts
@@ -1,0 +1,7 @@
+import { WalletError } from './WalletError'
+
+export class WalletImportPathExistsError extends WalletError {
+  public constructor(message: string, { cause }: { cause?: Error } = {}) {
+    super(message, { cause })
+  }
+}